+++ src/main/java/nextstep/ladder/domain/Point.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.domain;
 
 public class Point {
