+++ src/main/java/nextstep/ladder/domain/PointFactory.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.domain;
 
 import java.util.ArrayList;
 import java.util.Arrays;
