+++ src/main/java/nextstep/ladder/domain/PointGenerator.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.domain;
 
 public interface PointGenerator {
     Point nextPoint(Point current);
