+++ src/main/java/nextstep/ladder/domain/RandomPointGenerator.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.domain;
 
 import java.util.Random;
